Proper movement

+++ Vec3d.h
@@ -12,15 +12,15 @@
 
 class Vec3d {
 public:
-	Vec3d(float x, float y, float z)
+	Vec3d(double x, double y, double z)
 	{
 		this->x = x;
 		this->y = y;
 		this->z = z;
 	}
-	float x, y, z;
+	double x, y, z;
 	virtual ~Vec3d();
-	virtual float magnitude();
+	virtual double magnitude();
 	virtual Vec3d normalize();
 	virtual Vec3d operator + (const Vec3d);
 	virtual Vec3d operator - (const Vec3d);
@@ -36,14 +36,14 @@ Vec3d Vec3d::operator - (const Vec3d param)
 	return Vec3d(this->x - param.x, this->y - param.y, this->z - param.z);
 }
 
-float Vec3d::magnitude()
+double Vec3d::magnitude()
 {
 	return sqrt(pow(x, 2) + pow(y, 2) + pow(z, 2));
 }
 
 Vec3d Vec3d::normalize()
 {
-	float mag = magnitude();
+	double mag = magnitude();
 	if(mag == 0)mag = 0.0000001;
 	return Vec3d(x/mag, y/mag, z/mag);
 }
